+++ src/main/java/com/lckback/lckforall/aboutlck/controller/AboutLckMatchController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.lckback.lckforall.aboutlck.dto.FindMatchesByDateDto;
+import com.lckback.lckforall.aboutlck.dto.match.FindMatchesByDateDto;
 import com.lckback.lckforall.aboutlck.service.AboutLckMatchService;
 import com.lckback.lckforall.base.api.ApiResponse;
