+++ src/main/java/com/lckback/lckforall/aboutlck/converter/match/AboutMatchConverter.java
@@ -1,8 +1,8 @@
-package com.lckback.lckforall.aboutlck.converter;
+package com.lckback.lckforall.aboutlck.converter.match;
 
 import java.util.List;
 
-import com.lckback.lckforall.aboutlck.dto.FindMatchesByDateDto;
+import com.lckback.lckforall.aboutlck.dto.match.FindMatchesByDateDto;
 import com.lckback.lckforall.base.type.MatchResult;
 import com.lckback.lckforall.match.model.Match;
 import com.lckback.lckforall.team.model.Team;
