+++ src/main/java/com/lckback/lckforall/aboutlck/converter/team/FindTeamRatingHistoryConverter.java
@@ -1,8 +1,8 @@
-package com.lckback.lckforall.aboutlck.converter;
+package com.lckback.lckforall.aboutlck.converter.team;
 
 import org.springframework.data.domain.Page;
 
-import com.lckback.lckforall.aboutlck.dto.FindTeamRatingHistoryDto;
+import com.lckback.lckforall.aboutlck.dto.team.FindTeamRatingHistoryDto;
 import com.lckback.lckforall.team.model.SeasonTeam;
 
 public class FindTeamRatingHistoryConverter {
