+++ src/main/java/com/lckback/lckforall/aboutlck/repository/SeasonTeamRepository.java
@@ -1,10 +1,14 @@
 package com.lckback.lckforall.aboutlck.repository;
 
+import java.util.List;
+
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 
+import com.lckback.lckforall.player.model.SeasonTeamPlayer;
 import com.lckback.lckforall.team.model.Season;
 import com.lckback.lckforall.team.model.SeasonTeam;
 import com.lckback.lckforall.team.model.Team;
@@ -18,4 +22,13 @@ public interface SeasonTeamRepository extends JpaRepository<SeasonTeam, Long> {
 
 	@EntityGraph(attributePaths = {"season"})
 	Page<SeasonTeam> findAllByTeam(Team team, Pageable pageable);
+
+	@EntityGraph(attributePaths = {"team", "season"})
+	@Query("select st from SeasonTeam st join st.seasonTeamPlayers stp where stp in :seasonTeamPlayers")
+	Page<SeasonTeam> findAllBySeasonTeamPlayersIn(List<SeasonTeamPlayer> seasonTeamPlayers, Pageable pageable);
+
+	@EntityGraph(attributePaths = {"season"})
+	@Query("select st from SeasonTeam st join st.seasonTeamPlayers stp where st.rating = 1 and stp in :seasonTeamPlayers")
+	Page<SeasonTeam> findWinningSeasonTeamBySeasonTeamPlayers(List<SeasonTeamPlayer> seasonTeamPlayers,
+		Pageable pageable);
 }
