+++ src/main/java/com/lckback/lckforall/aboutlck/service/AboutLckMatchService.java
@@ -8,8 +8,8 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.lckback.lckforall.aboutlck.converter.AboutMatchConverter;
-import com.lckback.lckforall.aboutlck.dto.FindMatchesByDateDto;
+import com.lckback.lckforall.aboutlck.converter.match.AboutMatchConverter;
+import com.lckback.lckforall.aboutlck.dto.match.FindMatchesByDateDto;
 import com.lckback.lckforall.aboutlck.repository.MatchRepository;
 import com.lckback.lckforall.match.model.Match;
