+++ src/main/java/com/lckback/lckforall/aboutlck/service/AboutLckTeamService.java
@@ -7,14 +7,14 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.lckback.lckforall.aboutlck.converter.FindTeamPlayerHistoryConverter;
-import com.lckback.lckforall.aboutlck.converter.FindTeamRatingBySeasonConverter;
-import com.lckback.lckforall.aboutlck.converter.FindTeamRatingHistoryConverter;
-import com.lckback.lckforall.aboutlck.converter.FindTeamWinningHistoryConverter;
-import com.lckback.lckforall.aboutlck.dto.FindTeamPlayerHistoryDto;
-import com.lckback.lckforall.aboutlck.dto.FindTeamRatingHistoryDto;
-import com.lckback.lckforall.aboutlck.dto.FindTeamRatingBySeasonDto;
-import com.lckback.lckforall.aboutlck.dto.FindTeamWinningHistoryDto;
+import com.lckback.lckforall.aboutlck.converter.team.FindTeamPlayerHistoryConverter;
+import com.lckback.lckforall.aboutlck.converter.team.FindTeamRatingBySeasonConverter;
+import com.lckback.lckforall.aboutlck.converter.team.FindTeamRatingHistoryConverter;
+import com.lckback.lckforall.aboutlck.converter.team.FindTeamWinningHistoryConverter;
+import com.lckback.lckforall.aboutlck.dto.team.FindTeamPlayerHistoryDto;
+import com.lckback.lckforall.aboutlck.dto.team.FindTeamRatingHistoryDto;
+import com.lckback.lckforall.aboutlck.dto.team.FindTeamRatingBySeasonDto;
+import com.lckback.lckforall.aboutlck.dto.team.FindTeamWinningHistoryDto;
 import com.lckback.lckforall.aboutlck.repository.SeasonRepository;
 import com.lckback.lckforall.aboutlck.repository.SeasonTeamRepository;
 import com.lckback.lckforall.aboutlck.repository.TeamRepository;
